requirements.txt
spacy
pdfminer.six
openai-whisper
streamlit
textblob


